refactor(structure): reorganize src directory into formaverter package
Moved source files from the 'src' directory to a new 'formaverter' package to improve project organization and modularity. Updated import statements in both source and test files to reflect the new package structure. This change enhances maintainability by clearly defining the scope of the formaverter module, making it easier to manage dependencies and future expansions.
