+++ src/formaverter/image_collector.py
@@ -4,7 +4,7 @@ This module provides functionality to collect images from a directory.
 
 import os
 from typing import List
-from src.image_converter import ImageConverter
+from .image_converter import ImageConverter
 
 def collect_images(input_dir: str) -> List[str]:
     """
