+++ tests/test_image_collector.py
@@ -5,8 +5,8 @@ import tempfile
 import shutil
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
 
-from image_collector import collect_images
-from image_converter import ImageConverter
+from formaverter.image_collector import collect_images
+from formaverter.image_converter import ImageConverter
 
 class TestImageCollector(unittest.TestCase):
