+++ tests/test_image_converter.py
@@ -6,7 +6,7 @@ import unittest
 import tempfile
 import shutil
 from PIL import Image
-from image_converter import ImageConverter, convert_image
+from formaverter.image_converter import ImageConverter, convert_image
 
 class TestImageConverter(unittest.TestCase):
